requirements.txt
beautifulsoup4==4.12.2
fastapi>=0.116.0
uvicorn[standard]>=0.34.0
python-multipart==0.0.20
starlette>=0.45.3
werkzeug>=3.0.0
python-json-logger==2.0.7
requests==2.32.4
psycopg[binary,pool]==3.2.4
./vendor/inferd-0.5.9-0_v0.5.9-py3-none-any.whl
pydantic==2.11.5
Authlib==1.3.2
httpx~=0.28.0
attrs>=21.3.0
python-dateutil>=2.8.0
pytest==8.3.2
modal==1.0.3
gunicorn==23.0.0
google-cloud-language==2.15.0
google-cloud-storage==2.19.0
google-cloud-vision==3.9.0
rank_bm25==0.2.2
time-machine==2.16.0
google-cloud-recaptcha-enterprise==1.27.0
python-dotenv==1.0.1
types-protobuf==5.29.1.20250208
types-requests==2.32.0.20240914
types-beautifulsoup4==4.12.0.20250204
mypy==1.15.0
ruff==0.11.7
puremagic==1.28.0
sqlalchemy==2.0.40
alembic==1.15.2
alembic-postgresql-enum==1.7.0
pydub==0.25.1
openai==1.107.2
pydantic-ai==1.0.18
pytest-postgresql==7.0.2
mcp==1.13.1
beaker-py==2.5.0
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-gcp-trace==1.10.0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-instrumentation-httpx==0.59b0
opentelemetry-instrumentation-requests==0.59b0
opentelemetry-instrumentation-sqlalchemy==0.59b0
opentelemetry-instrumentation-psycopg==0.59b0